+++ package.json
@@ -9,18 +9,18 @@
     "@monaco-editor/react": "^4.4.6",
     "@radix-ui/react-tabs": "^1.0.3",
     "client-zip": "^2.3.1",
-    "code-block-writer": "^11.0.3",
+    "code-block-writer": "^12.0.0",
     "esbuild-wasm": "^0.14.39",
-    "monaco-editor": "^0.36.1",
+    "monaco-editor": "^0.37.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "devDependencies": {
-    "@figma/plugin-typings": "^1.58.0",
+    "@figma/plugin-typings": "^1.62.0",
     "@types/react": "^17.0.45",
     "@types/react-dom": "^17.0.17",
     "esbuild": "^0.14.39",
     "typescript": "^4.9.5",
-    "typescript-json-schema": "^0.55.0"
+    "typescript-json-schema": "^0.56.0"
   }
 }
