+++ src/utils/esbuild.ts
@@ -1,4 +1,5 @@
 import {initialize, transform} from 'esbuild-wasm';
+
 import type {Settings} from 'types/settings';
 
 const wasmURL = 'https://unpkg.com/esbuild-wasm@0.14.39/esbuild.wasm';
